Massive refactoring in progress, camera doesn't work
Matrix math is probably broken right now.
Orthographic camera isn't even tested.
Project1 class is horribly broken and won't work properly with all the
refactoring.
EventQueue about to be refactored to use RxJava library with
Observables.
EventQueue is now used for talking to classes. For example,
GameApplicationDisplay pushes a Pair upon a resize event, or the
KeyboardManager pushes events upon key presses/releases. Convenience
functions are provided for listening to the keyboard/mouse, but the
raw EventQueue API can be used for anything.
Sadly, when using the EventQueue, raw Objects are inserted and pushed
around, so a cast has to be made when that Object reaches a listener.

GameApplicationDisplay is the only starting point now, as LWJGL 2.9 and
greater have to double-draw when using a Frame/JFrame/etc.
GameApplicationDisplay much cleaner and actually documented.

Many more things to be done!